Added pass and intercept plot to gradio UI
